requirements.txt
streamlit==1.33.0
azure-ai-formrecognizer==3.3.2
requests==2.31.0
python-dotenv==1.0.0